refactor: . push TaskLocation-related methods to ListItem
