+++ src/Task/ListItem.ts
@@ -1,6 +1,7 @@
+import type { TasksFile } from '../Scripting/TasksFile';
+import type { Task } from './Task';
 import type { TaskLocation } from './TaskLocation';
 import { TaskRegularExpressions } from './TaskRegularExpressions';
-import type { Task } from './Task';
 
 export class ListItem {
     // The original line read from file.
@@ -122,5 +123,41 @@ export class ListItem {
         }
 
         return list1.every((item, index) => item.identicalTo(list2[index]));
+    }
+
+    public get path(): string {
+        return this.taskLocation.path;
+    }
+
+    public get file(): TasksFile {
+        return this.taskLocation.tasksFile;
+    }
+
+    /**
+     * Return the name of the file containing the task, with the .md extension removed.
+     */
+    public get filename(): string | null {
+        const fileNameMatch = this.path.match(/([^/]+)\.md$/);
+        if (fileNameMatch !== null) {
+            return fileNameMatch[1];
+        } else {
+            return null;
+        }
+    }
+
+    get lineNumber(): number {
+        return this.taskLocation.lineNumber;
+    }
+
+    get sectionStart(): number {
+        return this.taskLocation.sectionStart;
+    }
+
+    get sectionIndex(): number {
+        return this.taskLocation.sectionIndex;
+    }
+
+    public get precedingHeader(): string | null {
+        return this.taskLocation.precedingHeader;
     }
 }
